requirements.txt
Flask
python-dotenv
numpy
joblib
google-generativeai
